Type of AspNetUser Id of .NET Identity was changed from string to long. There was added seed to migrations to insert admin User, Categories and Products to database. Models for database were changed: User and Product has long type for Id now. Fixed bug when Cart had two User Id (one was FK, another - part of PK, however it was expected to be only one attribute with both properties).
Functionality that was implemented:
Get all categories
Get all products / all products of specific category
Get specific product
Search for products by name / for products of specific category
Get all products from Cart
Add product to Cart
Change product amount in Cart

All products go with pagination as well as with soldAmount and flag which shows whether the product is in Cart for current User.

Product doesn't have picture, neither does Category. To come in next commits.

+++ Data/Models/Cart.cs
@@ -12,12 +12,12 @@ namespace BlackMarket_API.Data.Models
         [Key]
         [Column(Order = 0)]
         [DatabaseGenerated(DatabaseGeneratedOption.None)]
-        public int UserId { get; set; }
+        public long UserId { get; set; }
 
         [Key]
         [Column(Order = 1)]
         [DatabaseGenerated(DatabaseGeneratedOption.None)]
-        public int ProductId { get; set; }
+        public long ProductId { get; set; }
 
         public int Amount { get; set; }
